requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
google-generativeai==0.3.0
python-pptx==0.6.21
Pillow==9.5.0
python-magic==0.4.27
python-jose==3.3.0
gunicorn==21.2.0 